# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Omni Task Manager API
  description: |
    API for the world's best task manager.

    ## Authentication
    All authenticated endpoints require a JWT bearer token in the Authorization header.

    ## Rate Limiting
    - Auth endpoints: 10 requests per 15 minutes
    - Password reset: 3 requests per hour
    - Bulk operations: 10 requests per minute
    - General API: 100 requests per minute
  version: 1.0.0
servers:
  - url: /api
    description: API server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Label:
      type: object
      required: [name, workspace_id]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 255
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        workspace_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Routine:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        schedule:
          type: object
        workspace_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
...
paths:
...
  /labels:
    get:
      tags: [Labels]
      summary: List labels
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of labels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Label'

    post:
      tags: [Labels]
      summary: Create label
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                color:
                  type: string
                project_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Label created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label'
  /labels/{id}:
    put:
      tags: [Labels]
      summary: Update label
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                color:
                  type: string
      responses:
        '200':
          description: Label updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label'

    delete:
      tags: [Labels]
      summary: Delete label
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Label deleted

  # Schedule
...
  /routines:
    get:
      tags: [Routines]
      summary: List routines
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of routines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Routine'

    post:
      tags: [Routines]
      summary: Create routine
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Routine'
      responses:
        '201':
          description: Routine created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Routine'

  # Time Tracking
...